+++ apis/repository/v1alpha1/zz_remotemavenrepository_types.go
@@ -72,7 +72,7 @@ type RemoteMavenRepositoryInitParameters struct {
 	ArchiveBrowsingEnabled *bool `json:"archiveBrowsingEnabled,omitempty" tf:"archive_browsing_enabled,omitempty"`
 
 	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time, an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed offline.
-	AssumedOfflinePeriodSecs *float64 `json:"assumedOfflinePeriodSecs,omitempty" tf:"assumed_offline_period_secs,omitempty"`
+	AssumedOfflinePeriodSecs *int64 `json:"assumedOfflinePeriodSecs,omitempty" tf:"assumed_offline_period_secs,omitempty"`
 
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact resolution.
 	BlackedOut *bool `json:"blackedOut,omitempty" tf:"blacked_out,omitempty"`
@@ -137,16 +137,16 @@ type RemoteMavenRepositoryInitParameters struct {
 	LocalAddress *string `json:"localAddress,omitempty" tf:"local_address,omitempty"`
 
 	// The maximum number of unique snapshots of a single artifact to store. Once the number of snapshots exceeds this setting, older versions are removed. A value of 0 (default) indicates there is no limit, and unique snapshots are not cleaned up.
-	MaxUniqueSnapshots *float64 `json:"maxUniqueSnapshots,omitempty" tf:"max_unique_snapshots,omitempty"`
+	MaxUniqueSnapshots *int64 `json:"maxUniqueSnapshots,omitempty" tf:"max_unique_snapshots,omitempty"`
 
 	// Metadata Retrieval Cache Timeout (Sec) in the UI.This value refers to the number of seconds to wait for retrieval from the remote before serving locally cached artifact or fail the request.
-	MetadataRetrievalTimeoutSecs *float64 `json:"metadataRetrievalTimeoutSecs,omitempty" tf:"metadata_retrieval_timeout_secs,omitempty"`
+	MetadataRetrievalTimeoutSecs *int64 `json:"metadataRetrievalTimeoutSecs,omitempty" tf:"metadata_retrieval_timeout_secs,omitempty"`
 
 	// The set of mime types that should override the block_mismatching_mime_types setting. Eg: 'application/json,application/xml'. Default value is empty.
 	MismatchingMimeTypesOverrideList *string `json:"mismatchingMimeTypesOverrideList,omitempty" tf:"mismatching_mime_types_override_list,omitempty"`
 
 	// Missed Retrieval Cache Period (Sec) in the UI. The number of seconds to cache artifact retrieval misses (artifact not found). A value of 0 indicates no caching.
-	MissedCachePeriodSeconds *float64 `json:"missedCachePeriodSeconds,omitempty" tf:"missed_cache_period_seconds,omitempty"`
+	MissedCachePeriodSeconds *int64 `json:"missedCachePeriodSeconds,omitempty" tf:"missed_cache_period_seconds,omitempty"`
 
 	// Internal description.
 	Notes *string `json:"notes,omitempty" tf:"notes,omitempty"`
@@ -189,12 +189,12 @@ type RemoteMavenRepositoryInitParameters struct {
 	RepoLayoutRef *string `json:"repoLayoutRef,omitempty" tf:"repo_layout_ref,omitempty"`
 
 	// Metadata Retrieval Cache Period (Sec) in the UI. This value refers to the number of seconds to cache metadata files before checking for newer versions on remote server. A value of 0 indicates no caching.
-	RetrievalCachePeriodSeconds *float64 `json:"retrievalCachePeriodSeconds,omitempty" tf:"retrieval_cache_period_seconds,omitempty"`
+	RetrievalCachePeriodSeconds *int64 `json:"retrievalCachePeriodSeconds,omitempty" tf:"retrieval_cache_period_seconds,omitempty"`
 
 	ShareConfiguration *bool `json:"shareConfiguration,omitempty" tf:"share_configuration,omitempty"`
 
 	// Network timeout (in ms) to use when establishing a connection and for unanswered requests. Timing out on a network operation is considered a retrieval failure.
-	SocketTimeoutMillis *float64 `json:"socketTimeoutMillis,omitempty" tf:"socket_timeout_millis,omitempty"`
+	SocketTimeoutMillis *int64 `json:"socketTimeoutMillis,omitempty" tf:"socket_timeout_millis,omitempty"`
 
 	// When set, the repository should store cached artifacts locally. When not set, artifacts are not stored locally, and direct repository-to-client streaming is used. This can be useful for multi-server setups over a high-speed LAN, with one Artifactory caching certain data on central storage, and streaming it directly to satellite pass-though Artifactory servers.
 	StoreArtifactsLocally *bool `json:"storeArtifactsLocally,omitempty" tf:"store_artifacts_locally,omitempty"`
@@ -209,7 +209,7 @@ type RemoteMavenRepositoryInitParameters struct {
 	URL *string `json:"url,omitempty" tf:"url,omitempty"`
 
 	// Unused Artifacts Cleanup Period (Hr) in the UI. The number of hours to wait before an artifact is deemed 'unused' and eligible for cleanup from the repository. A value of 0 means automatic cleanup of cached artifacts is disabled.
-	UnusedArtifactsCleanupPeriodHours *float64 `json:"unusedArtifactsCleanupPeriodHours,omitempty" tf:"unused_artifacts_cleanup_period_hours,omitempty"`
+	UnusedArtifactsCleanupPeriodHours *int64 `json:"unusedArtifactsCleanupPeriodHours,omitempty" tf:"unused_artifacts_cleanup_period_hours,omitempty"`
 
 	Username *string `json:"username,omitempty" tf:"username,omitempty"`
 
@@ -227,7 +227,7 @@ type RemoteMavenRepositoryObservation struct {
 	ArchiveBrowsingEnabled *bool `json:"archiveBrowsingEnabled,omitempty" tf:"archive_browsing_enabled,omitempty"`
 
 	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time, an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed offline.
-	AssumedOfflinePeriodSecs *float64 `json:"assumedOfflinePeriodSecs,omitempty" tf:"assumed_offline_period_secs,omitempty"`
+	AssumedOfflinePeriodSecs *int64 `json:"assumedOfflinePeriodSecs,omitempty" tf:"assumed_offline_period_secs,omitempty"`
 
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact resolution.
 	BlackedOut *bool `json:"blackedOut,omitempty" tf:"blacked_out,omitempty"`
@@ -294,16 +294,16 @@ type RemoteMavenRepositoryObservation struct {
 	LocalAddress *string `json:"localAddress,omitempty" tf:"local_address,omitempty"`
 
 	// The maximum number of unique snapshots of a single artifact to store. Once the number of snapshots exceeds this setting, older versions are removed. A value of 0 (default) indicates there is no limit, and unique snapshots are not cleaned up.
-	MaxUniqueSnapshots *float64 `json:"maxUniqueSnapshots,omitempty" tf:"max_unique_snapshots,omitempty"`
+	MaxUniqueSnapshots *int64 `json:"maxUniqueSnapshots,omitempty" tf:"max_unique_snapshots,omitempty"`
 
 	// Metadata Retrieval Cache Timeout (Sec) in the UI.This value refers to the number of seconds to wait for retrieval from the remote before serving locally cached artifact or fail the request.
-	MetadataRetrievalTimeoutSecs *float64 `json:"metadataRetrievalTimeoutSecs,omitempty" tf:"metadata_retrieval_timeout_secs,omitempty"`
+	MetadataRetrievalTimeoutSecs *int64 `json:"metadataRetrievalTimeoutSecs,omitempty" tf:"metadata_retrieval_timeout_secs,omitempty"`
 
 	// The set of mime types that should override the block_mismatching_mime_types setting. Eg: 'application/json,application/xml'. Default value is empty.
 	MismatchingMimeTypesOverrideList *string `json:"mismatchingMimeTypesOverrideList,omitempty" tf:"mismatching_mime_types_override_list,omitempty"`
 
 	// Missed Retrieval Cache Period (Sec) in the UI. The number of seconds to cache artifact retrieval misses (artifact not found). A value of 0 indicates no caching.
-	MissedCachePeriodSeconds *float64 `json:"missedCachePeriodSeconds,omitempty" tf:"missed_cache_period_seconds,omitempty"`
+	MissedCachePeriodSeconds *int64 `json:"missedCachePeriodSeconds,omitempty" tf:"missed_cache_period_seconds,omitempty"`
 
 	// Internal description.
 	Notes *string `json:"notes,omitempty" tf:"notes,omitempty"`
@@ -346,12 +346,12 @@ type RemoteMavenRepositoryObservation struct {
 	RepoLayoutRef *string `json:"repoLayoutRef,omitempty" tf:"repo_layout_ref,omitempty"`
 
 	// Metadata Retrieval Cache Period (Sec) in the UI. This value refers to the number of seconds to cache metadata files before checking for newer versions on remote server. A value of 0 indicates no caching.
-	RetrievalCachePeriodSeconds *float64 `json:"retrievalCachePeriodSeconds,omitempty" tf:"retrieval_cache_period_seconds,omitempty"`
+	RetrievalCachePeriodSeconds *int64 `json:"retrievalCachePeriodSeconds,omitempty" tf:"retrieval_cache_period_seconds,omitempty"`
 
 	ShareConfiguration *bool `json:"shareConfiguration,omitempty" tf:"share_configuration,omitempty"`
 
 	// Network timeout (in ms) to use when establishing a connection and for unanswered requests. Timing out on a network operation is considered a retrieval failure.
-	SocketTimeoutMillis *float64 `json:"socketTimeoutMillis,omitempty" tf:"socket_timeout_millis,omitempty"`
+	SocketTimeoutMillis *int64 `json:"socketTimeoutMillis,omitempty" tf:"socket_timeout_millis,omitempty"`
 
 	// When set, the repository should store cached artifacts locally. When not set, artifacts are not stored locally, and direct repository-to-client streaming is used. This can be useful for multi-server setups over a high-speed LAN, with one Artifactory caching certain data on central storage, and streaming it directly to satellite pass-though Artifactory servers.
 	StoreArtifactsLocally *bool `json:"storeArtifactsLocally,omitempty" tf:"store_artifacts_locally,omitempty"`
@@ -366,7 +366,7 @@ type RemoteMavenRepositoryObservation struct {
 	URL *string `json:"url,omitempty" tf:"url,omitempty"`
 
 	// Unused Artifacts Cleanup Period (Hr) in the UI. The number of hours to wait before an artifact is deemed 'unused' and eligible for cleanup from the repository. A value of 0 means automatic cleanup of cached artifacts is disabled.
-	UnusedArtifactsCleanupPeriodHours *float64 `json:"unusedArtifactsCleanupPeriodHours,omitempty" tf:"unused_artifacts_cleanup_period_hours,omitempty"`
+	UnusedArtifactsCleanupPeriodHours *int64 `json:"unusedArtifactsCleanupPeriodHours,omitempty" tf:"unused_artifacts_cleanup_period_hours,omitempty"`
 
 	Username *string `json:"username,omitempty" tf:"username,omitempty"`
 
@@ -387,7 +387,7 @@ type RemoteMavenRepositoryParameters struct {
 
 	// The number of seconds the repository stays in assumed offline state after a connection error. At the end of this time, an online check is attempted in order to reset the offline status. A value of 0 means the repository is never assumed offline.
 	// +kubebuilder:validation:Optional
-	AssumedOfflinePeriodSecs *float64 `json:"assumedOfflinePeriodSecs,omitempty" tf:"assumed_offline_period_secs,omitempty"`
+	AssumedOfflinePeriodSecs *int64 `json:"assumedOfflinePeriodSecs,omitempty" tf:"assumed_offline_period_secs,omitempty"`
 
 	// (A.K.A 'Ignore Repository' on the UI) When set, the repository or its local cache do not participate in artifact resolution.
 	// +kubebuilder:validation:Optional
@@ -474,11 +474,11 @@ type RemoteMavenRepositoryParameters struct {
 
 	// The maximum number of unique snapshots of a single artifact to store. Once the number of snapshots exceeds this setting, older versions are removed. A value of 0 (default) indicates there is no limit, and unique snapshots are not cleaned up.
 	// +kubebuilder:validation:Optional
-	MaxUniqueSnapshots *float64 `json:"maxUniqueSnapshots,omitempty" tf:"max_unique_snapshots,omitempty"`
+	MaxUniqueSnapshots *int64 `json:"maxUniqueSnapshots,omitempty" tf:"max_unique_snapshots,omitempty"`
 
 	// Metadata Retrieval Cache Timeout (Sec) in the UI.This value refers to the number of seconds to wait for retrieval from the remote before serving locally cached artifact or fail the request.
 	// +kubebuilder:validation:Optional
-	MetadataRetrievalTimeoutSecs *float64 `json:"metadataRetrievalTimeoutSecs,omitempty" tf:"metadata_retrieval_timeout_secs,omitempty"`
+	MetadataRetrievalTimeoutSecs *int64 `json:"metadataRetrievalTimeoutSecs,omitempty" tf:"metadata_retrieval_timeout_secs,omitempty"`
 
 	// The set of mime types that should override the block_mismatching_mime_types setting. Eg: 'application/json,application/xml'. Default value is empty.
 	// +kubebuilder:validation:Optional
@@ -486,7 +486,7 @@ type RemoteMavenRepositoryParameters struct {
 
 	// Missed Retrieval Cache Period (Sec) in the UI. The number of seconds to cache artifact retrieval misses (artifact not found). A value of 0 indicates no caching.
 	// +kubebuilder:validation:Optional
-	MissedCachePeriodSeconds *float64 `json:"missedCachePeriodSeconds,omitempty" tf:"missed_cache_period_seconds,omitempty"`
+	MissedCachePeriodSeconds *int64 `json:"missedCachePeriodSeconds,omitempty" tf:"missed_cache_period_seconds,omitempty"`
 
 	// Internal description.
 	// +kubebuilder:validation:Optional
@@ -543,14 +543,14 @@ type RemoteMavenRepositoryParameters struct {
 
 	// Metadata Retrieval Cache Period (Sec) in the UI. This value refers to the number of seconds to cache metadata files before checking for newer versions on remote server. A value of 0 indicates no caching.
 	// +kubebuilder:validation:Optional
-	RetrievalCachePeriodSeconds *float64 `json:"retrievalCachePeriodSeconds,omitempty" tf:"retrieval_cache_period_seconds,omitempty"`
+	RetrievalCachePeriodSeconds *int64 `json:"retrievalCachePeriodSeconds,omitempty" tf:"retrieval_cache_period_seconds,omitempty"`
 
 	// +kubebuilder:validation:Optional
 	ShareConfiguration *bool `json:"shareConfiguration,omitempty" tf:"share_configuration,omitempty"`
 
 	// Network timeout (in ms) to use when establishing a connection and for unanswered requests. Timing out on a network operation is considered a retrieval failure.
 	// +kubebuilder:validation:Optional
-	SocketTimeoutMillis *float64 `json:"socketTimeoutMillis,omitempty" tf:"socket_timeout_millis,omitempty"`
+	SocketTimeoutMillis *int64 `json:"socketTimeoutMillis,omitempty" tf:"socket_timeout_millis,omitempty"`
 
 	// When set, the repository should store cached artifacts locally. When not set, artifacts are not stored locally, and direct repository-to-client streaming is used. This can be useful for multi-server setups over a high-speed LAN, with one Artifactory caching certain data on central storage, and streaming it directly to satellite pass-though Artifactory servers.
 	// +kubebuilder:validation:Optional
@@ -570,7 +570,7 @@ type RemoteMavenRepositoryParameters struct {
 
 	// Unused Artifacts Cleanup Period (Hr) in the UI. The number of hours to wait before an artifact is deemed 'unused' and eligible for cleanup from the repository. A value of 0 means automatic cleanup of cached artifacts is disabled.
 	// +kubebuilder:validation:Optional
-	UnusedArtifactsCleanupPeriodHours *float64 `json:"unusedArtifactsCleanupPeriodHours,omitempty" tf:"unused_artifacts_cleanup_period_hours,omitempty"`
+	UnusedArtifactsCleanupPeriodHours *int64 `json:"unusedArtifactsCleanupPeriodHours,omitempty" tf:"unused_artifacts_cleanup_period_hours,omitempty"`
 
 	// +kubebuilder:validation:Optional
 	Username *string `json:"username,omitempty" tf:"username,omitempty"`
@@ -612,7 +612,7 @@ type RemoteMavenRepositoryStatus struct {
 // +kubebuilder:printcolumn:name="READY",type="string",JSONPath=".status.conditions[?(@.type=='Ready')].status"
 // +kubebuilder:printcolumn:name="EXTERNAL-NAME",type="string",JSONPath=".metadata.annotations.crossplane\\.io/external-name"
 // +kubebuilder:printcolumn:name="AGE",type="date",JSONPath=".metadata.creationTimestamp"
-// +kubebuilder:resource:scope=Cluster,categories={crossplane,managed,jfrogartifactory}
+// +kubebuilder:resource:scope=Cluster,categories={crossplane,managed,artifactory}
 type RemoteMavenRepository struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
